Merge Measure Bug Fix Validation Tests (#339)
What does this Pull Request accomplish?

Added auto tests that validate the Merge Measure Bug Fix. The Bug is that when pins are merged, the measurements that return the `PinSiteData` object must associate the result with the group name instead of the pin name.
Why should this Pull Request be merged?

Additional auto tests for the Merge feature Measurement Bug. This test covers all the user workflow APIs.
What testing has been done?

Tested locally on a VM with auto tests